Fix for :
Error occurs when there is a security group named "A" in EntraID, and there is also a group named "A" in the Databricks account console. If the "A" group in EntraID is empty, but the "A" group in the Databricks account console contains users, running the relevant program will trigger the error.
Error message is like this:

TypeError: 'NoneType' object is not iterable